requirements.txt
Flask>=1.0.2
requests>=2.20.1
PyYAML>=5.1.1
beautifulsoup4
lxml